Add files via upload
l. The properties of the Node are defined in the Block Class and is mostly composed of strings. A Blockchain Class is developed for utility functions like adding and getting blocks, getting the height of the blockchain, and validating the chain. Validating is done by comparing the hashes of the adjacent blocks and seeing if it matches. Time and space complexities of singly linked list applies. Adding blocks and getting the height are in constant time while searching and validating will be O(n), where n is the number of nodes we have to traverse. Space complexity is also O(n) dependent on the said number of nodes. Extra space occupied will grow linearly as we add more blocks to the chain.
"

- The Logic is to write similar to a Single Linked List
- A genesis block is created as the head and all succeeding blocks are appended at the    tail
- But instead of next node, we link the previous node of each node
- Here, the link to the previous node is the Previous Hash
- Each, Block of the Block Chain is like a Node of a Linked List
- Validation is done by comparing the hashes of the adjacent blocks and seeing if it matches

Choice of Data Structures id Hash Map and Single Linked List

Complexity:-
	- Since, the Logic is implemented using a concept of Linked List
        - Each element of Linked List is a Hash
        - As the Linked List will consist of n number of nodes, hence we need to traverse n nodes
        - Hence, the time complexity is O(n)

Space Complexity :- 
         - Logic is implemented by using a Linked List
	 - Hence, for the complexities, rules of Linked List will be applicable
         - If we have n number of nodes, the Space Complexity will be O(n)
	 - Space Complexity will increase if more blocks are added